backend/requirements.txt
openai
sqlalchemy
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
passlib[bcrypt]
bcrypt==4.3.0
pydantic[email]
pyjwt[crypto]
